fix the quantize_image function of image.c
There is no return value in the quantize_image of image.c. After setting BUILD_SHARED_LIBS and using cmake to build .sln, we open it and generate the ALL_BUILD project. But there is an error showing that quantize_image has no return value. So I add the return value in the function.