Update darknet_video.py
I was getting a memory problem because as `darknet_image` is created in every iteration of `video_capture`, it is not being released by the `darknet_image_queue` when it gets `.get()`. 

To fix it I make a list of darknet_images that is larger than the size of the queue. The list is being repeatedly updated with new darknet_images in `video_capture`, but the image that has not been processed yet, is never overwritten. A bit of a hacky fix, but it works.